[feat #109] 리뷰페이지 리팩토링 완료
