+++ src/components/reviewComponents/ReviewList.jsx
@@ -1,4 +1,3 @@
-import { supabase } from 'api/supabase/supabase';
 import defaultProfileImage from 'assets/defaultProfileImage.png';
 import { getFormattedDate } from 'components/communityComponents/formattedDate';
 import {
@@ -12,42 +11,28 @@ import {
   ListNickName,
   ToggleContent
 } from 'components/styles/ReviewStyle';
+import useSetMutation from 'hooks/useSetMutations';
 import { useState } from 'react';
+import { useQuery } from 'react-query';
 import { ReviewUpdateForm } from './ReviewUpdateForm';
+import { deleteReview, getReviewList } from './reviewPageSupabase';
 
-export const ReviewList = ({ reviewData, setReviewData, placename, userData }) => {
+export const ReviewList = ({ placename, userData }) => {
+  const [mutaion] = useSetMutation(deleteReview, 'reviewList');
+
   // 수정 여부 state
   const [editDataId, setEditDataId] = useState(null);
   const [openItemId, setOpenItemId] = useState(null);
 
+  const { isLoading, data } = useQuery('reviewList', getReviewList);
+
+  if (isLoading) {
+    return <div>로딩중입니다</div>;
+  }
   // 데이터 삭제
   const removeReview = async (id, reviewimg) => {
     if (window.confirm('게시물을 삭제하시겠습니까?')) {
-      // 이미지 삭제
-      try {
-        const { data, error } = await supabase.storage.from('reviewImage').remove(reviewimg);
-        if (!error) {
-          console.log('이미지 삭제 성공', data);
-        }
-      } catch (error) {
-        console.log('이미지 삭제 실패', error);
-      }
-      // 게시글 삭제
-      try {
-        const { error } = await supabase.from('reviewWrite').delete().eq('id', id);
-        if (!error) {
-          alert('게시물이 삭제되었습니다.');
-          console.log('게시물 삭제 성공');
-          setReviewData((prev) =>
-            prev.filter((item) => {
-              return item.id !== id;
-            })
-          );
-          return;
-        }
-      } catch (error) {
-        console.log('게시물 삭제 실패', error);
-      }
+      mutaion.mutate([id, reviewimg]);
     }
   };
 
@@ -58,13 +43,13 @@ export const ReviewList = ({ reviewData, setReviewData, placename, userData }) =
 
   return (
     <ContentsList>
-      {reviewData
+      {data
         .sort((a, b) => new Date(b.date) - new Date(a.date))
         ?.map((item) => {
           return placename.placename === item.marker ? (
             <ListMapWrapper key={item.id}>
               {editDataId === item.id ? (
-                <ReviewUpdateForm item={item} setEditDataId={setEditDataId} setReviewData={setReviewData} />
+                <ReviewUpdateForm item={item} setEditDataId={setEditDataId} />
               ) : (
                 <>
                   {openItemId === item.id ? (
@@ -80,22 +65,26 @@ export const ReviewList = ({ reviewData, setReviewData, placename, userData }) =
                       </ContentImgWrapper>
                       <ContentBtnsWrapper>
                         <ListContent>{item.content}</ListContent>
-                        <ContentBtns>
-                          <button
-                            onClick={() => {
-                              setEditDataId(item.id);
-                            }}
-                          >
-                            수정
-                          </button>
-                          <button
-                            onClick={() => {
-                              removeReview(item.id, item.reviewimg);
-                            }}
-                          >
-                            삭제
-                          </button>
-                        </ContentBtns>
+                        {userData.email === item.email ? (
+                          <ContentBtns>
+                            <button
+                              onClick={() => {
+                                setEditDataId(item.id);
+                              }}
+                            >
+                              수정
+                            </button>
+                            <button
+                              onClick={() => {
+                                removeReview(item.id, item.reviewimg);
+                              }}
+                            >
+                              삭제
+                            </button>
+                          </ContentBtns>
+                        ) : (
+                          false
+                        )}
                       </ContentBtnsWrapper>
                     </ContentWrapper>
                   ) : (
