+++ src/pages/ReviewPage.jsx
@@ -12,31 +12,11 @@ import { useNavigate, useParams } from 'react-router-dom';
 
 export const ReviewPage = () => {
   // 데이터베이스에 저장된 게시물을 담는 state
-  const [reviewData, setReviewData] = useState([]);
   const placename = useParams();
   const { isLoading, data: userData } = useQuery('userData', getUser, {
     refetchOnWindowFocus: false
   });
   const navigate = useNavigate();
-
-  useEffect(() => {
-    // 게시글 불러오기
-    const fetchData = async () => {
-      let { data: reviewWrite, error } = await supabase.from('reviewWrite').select('*');
-      if (error) {
-        console.log('게시물 조회 실패', error);
-      } else {
-        // 이미지 불러오기
-        const reviewsWriteData = reviewWrite.map((item) => {
-          const imgUrl = supabase.storage.from('reviewImage').getPublicUrl(item.reviewimg);
-          return { ...item, imageUrl: imgUrl.data.publicUrl };
-        });
-        setReviewData(reviewsWriteData);
-      }
-    };
-
-    fetchData();
-  }, []);
 
   if (isLoading) {
     return <Loading />;
@@ -54,8 +34,8 @@ export const ReviewPage = () => {
     <>
       <ReviewHeader placename={placename} />
       <ReviewContainer>
-        <ReviewForm setReviewData={setReviewData} placename={placename} />
-        <ReviewList reviewData={reviewData} setReviewData={setReviewData} placename={placename} userData={userData} />
+        <ReviewForm placename={placename} />
+        <ReviewList placename={placename} userData={userData} />
       </ReviewContainer>
     </>
   );
